Scripts structure for camera reworked. toStatusName(status) separated into single cpp with header. CMakeLists changed accordingly
