+++ packages/camera/include/camera_status.h
@@ -1,0 +1,5 @@
+#include <string_view>
+
+std::string_view toStatusName(int status);
+
+int len(const std::string_view& s);
