+++ packages/camera/scripts/camera_identifier/CMakeLists.txt
@@ -6,5 +6,9 @@ add_compile_options(-Wall -Wextra -Wpedantic -Wno-dev)
 add_library(mvsdk SHARED IMPORTED)
 set_target_properties(mvsdk PROPERTIES IMPORTED_LOCATION "/lib/libMVSDK.so")
 
-add_executable(camera_identifier src/identifier.cpp)
+add_executable(camera_identifier src/identifier.cpp ../../src/camera_status.cpp)
+target_include_directories(camera_identifier PUBLIC
+  "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../include>"
+  "/usr/include")
+
 target_link_libraries(camera_identifier mvsdk)
