+++ packages/camera/scripts/camera_identifier/execute.sh
